requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
alembic==1.4.2
altgraph==0.17
ansicolors==1.1.8
ansiwrap==0.8.4
asgiref==3.2.5
asn1crypto==0.24.0
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
auto-py-to-exe==2.7.8
autobahn==20.7.1
Automat==0.6.0
backcall==0.2.0
beautifulsoup4==4.8.2
beautifultable==1.0.1
bigfloat==0.4.0
bleach==3.1.5
blinker==1.4
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2019.11.28
cffi==1.14.0
chainmap==1.0.3
chardet==3.0.4
click==7.1.2
cloud-init==20.3
colorama==0.3.7
combomethod==1.0.12
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
constants==0.6.0
covid==2.3.0
crypto==1.4.1
cryptography==3.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.4
Eel==0.12.4
entrypoints==0.3
enum34==1.1.10
evdev==1.4.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flaskwebgui==0.1.9
future==0.18.2
gast==0.3.2
gevent==20.6.2
gevent-websocket==0.10.1
google-pasta==0.1.8
graphviz==0.13.2
greenlet==0.4.16
grpcio==1.26.0
h5py==2.10.0
hashbang==0.1.14
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
idna-ssl==1.1.0
importlib-metadata==3.4.0
incremental==16.10.1
ipaddress==1.0.23
ipykernel==5.3.0
ipyparallel==6.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.14.1
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server-proxy==1.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
lxml==4.5.0
Mako==1.1.3
manage.py==0.2.10
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mementos==1.3.1
mistune==0.8.4
mongo==0.2.0
multidict==4.7.6
mysql-connector-python==8.0.20
Naked==0.1.31
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
nltk==3.5
notebook==6.0.3
nulltype==2.3.1
numpy==1.18.5
oauthlib==2.0.6
olefile==0.45.1
options==1.4.10
packaging==20.4
PAM==0.4.2
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plyer==1.4.3
prettytable==0.7.2
progressbar2==3.50.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
public==2019.4.13
pyang==2.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pydantic==1.6.1
pydotplus==2.0.2
pygame==1.9.6
Pygments==2.6.1
pygobject==3.26.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.7
PyJWT==1.5.3
pymongo==3.11.0
PyMySQL==0.9.3
pynput==1.7.3
pyOpenSSL==17.5.0
pyparsing==2.4.6
pyprog==1.1.0.post0
pyserial==3.4
PySimpleGUI==4.29.0.15
python-apt==1.6.5+ubuntu0.4
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
python-utils==2.4.0
python-xlib==0.29
pytz==2020.1
pyxdg==0.25
PyYAML==5.3
pyzmq==19.0.1
pyzo==4.7.4
qrcode==6.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.7.14
requests==2.23.0
requests-unixsocket==0.1.5
rsa==4.6
say==1.6.6
scikit-learn==0.22.1
scipy==1.4.1
screeninfo==0.6.5
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==16.0.0
shebang==2019.4.13
shellescape==3.8.1
simpervisor==0.3
simple-term-menu==0.4.1
simplere==1.2.13
six==1.15.0
sklearn==0.0
sockets==1.0.0
soupsieve==2.0
speedtest-cli==2.1.1
SQLAlchemy==1.3.18
sqlparse==0.3.1
ssh-import-id==5.7
swampy==3.0.1
systemd-python==234
tabulate==0.8.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.3
textwrap3==0.9.2
tornado==6.0.4
tqdm==4.41.1
traitlets==4.3.3
Twisted==17.9.0
txaio==20.4.1
typing-extensions==3.7.4.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vpython==7.6.1
walkdir==0.4.1
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
win-nic==2.0.1
words==0.0.1
wrapt==1.11.2
wsl-path-converter==0.3.1
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.2.8
yarl==1.4.2
zipp==3.4.0
zope.event==4.4
zope.interface==4.3.2
